requirements.txt
colorama
droidrun[google,anthropic,openai,deepseek,ollama,dev]
llama-index-llms-google-genai
python-dotenv
rich
